# Reintroduce Requires.jl since it seems to be faster now