--- a/src/glob.jl
+++ b/src/glob.jl
@@ -1,3 +1,6 @@
+using .Glob
+using .Glob: GlobMatch
+
 Base.readdir(pattern::GlobMatch, prefix::AbstractPath) = glob(pattern, prefix)
 
 function Glob.glob(pattern, prefix::T) where T<:AbstractPath