AFAICT, some of the performance issues with using Requires.jl are less relevant on Julia 1.5? Since we've already separated out the FilePathsBase.jl logic into a separate package this seems like a safer way to move forward.

Before re-introducing Requires.jl:
```julia
julia> @time using FilePaths
  0.161193 seconds (245.47 k allocations: 16.156 MiB)

julia> @time using FilePaths
  0.774819 seconds (1.74 M allocations: 87.113 MiB, 2.66% gc time)

julia> @time using FilePaths
  0.000248 seconds (280 allocations: 17.531 KiB)

julia> @time using FilePaths
  0.000251 seconds (280 allocations: 17.531 KiB)
```
Load times after switching to Requires.jl
```julia
julia> @time using FilePaths
  0.282045 seconds (386.24 k allocations: 23.427 MiB)

julia> @time using FilePaths
  0.785730 seconds (1.74 M allocations: 87.117 MiB, 1.33% gc time)

julia> @time using FilePaths
  0.000237 seconds (280 allocations: 17.531 KiB)

julia> @time using FilePaths
  0.000260 seconds (280 allocations: 17.531 KiB)

julia> @time using FilePaths
  0.000276 seconds (280 allocations: 17.531 KiB)
```
Same seems to go if our optional dep is loaded.

Before:
```julia
julia> @time using Glob, FilePaths
  0.188511 seconds (269.56 k allocations: 17.710 MiB)

julia> @time using Glob, FilePaths
  0.788237 seconds (1.74 M allocations: 87.125 MiB, 2.77% gc time)

julia> @time using Glob, FilePaths
  0.000393 seconds (577 allocations: 36.141 KiB)

julia> @time using Glob, FilePaths
  0.000416 seconds (577 allocations: 36.141 KiB)
```
After:
```julia
julia> @time using Glob, FilePaths
  1.192447 seconds (2.02 M allocations: 105.242 MiB, 10.89% gc time)

julia> @time using Glob, FilePaths
  0.085030 seconds (135.14 k allocations: 6.965 MiB)

julia> @time using Glob, FilePaths
  0.000441 seconds (577 allocations: 36.141 KiB)

julia> @time using Glob, FilePaths
  0.000387 seconds (577 allocations: 36.141 KiB)
```
Original performance issues reported [here](https://github.com/JuliaPackaging/Requires.jl/issues/39)