@oxinabox 2021-02-21
Don't you need to also time  how long it then takes to do `using Glob`? Etc
I thought the main slowness of Revise was in how much it increased time after loading the conditional.


Also deleting things from the dependency list is giving up on letting the package manager do its job and enforcing SemVer.
So while you might want to do this, I think claiming be this is

> a safer way to move forward.

Is odd.
Maybe I misunderstood which thing is safer?
Or do you mean: "we are doing this unsafe thing, but at least we are not doing it to FilePathsBase.jl"