@rofinn 2021-02-22
> Don't you need to also time how long it then takes to do using Glob? Etc

Yes, but the same seems to apply there as well. I've update my description with those times as well. NOTE: I'm only testing on 1.5.

> I thought the main slowness of Revise was in how much it increased time after loading the conditional.

Before it was recommended to put the `@require` call the primary slow was constantly re-triggering precompilation. There is a slow down in that the glue code isn't precompiled, but in our case that's quite small for now. I think the work around for that is to make the glue code its own package?

> Also deleting things from the dependency list is giving up on letting the package manager do its job and enforcing SemVer.

Agreed, but since the package manager doesn't support optional dependencies I don't see a good solution.

1. As I try to support more of the ecosystem I don't want to introduce a massive list of required dependencies
2. I don't want a glue package for every extension that'll likely be 1 or 2 simple method overrides

> Maybe I misunderstood which thing is safer?
> Or do you mean: "we are doing this unsafe thing, but at least we are not doing it to FilePathsBase.jl"

Yeah, I was thinking of compared to when I tried this with the FilePathsBase.jl code. I think the goal is for this to be a repository of simple glue bits which should be small enough to be embedded in projects if they have strict package dependency requirements.